Update Rsync.pm
Fix issue where no files are restored to a different location:

From BackupPC-devel 2015-05-13:
"It seems that in $installdir/lib/BackupPC/Xfer/Rsync.pm, $srcDir is set to 
"/", never changed, and passed as an argument to rsync_bpc. Yet the 
filenames in the "files-from" file are modified to strip any parent 
directories. Therefore $srcDir needs a different value if the restored 
files' paths aren't absolute relative to the root of the backup.

The included patch (inline and attached) fixes this problem for me; I've 
performed a few restores to original and different destinations and it 
appears to work as expected."